WIP updating the remaining time on a task's display
Current progress has a method that will update the display accurately when clicked, so next is to make this happen automatically every X seconds.
Changed the Task init to use a Double instead of an Int.
Added testing button that removes tasks from the list for testing the updating display