Views and navigation for account creation implemented
Created the Accunt sign in and creation views.
Account creation, login, and logout are still WIP.
Added Firebase Auth and Core packages to the project.
AuthenticationModel is a leftover from a previous tutorial attempt, I'll try to refactor it to be useful but it might be deleted in the future.
Added a new item to the AppPage enum.
Added some Firebase registration code to the MHDungeonApp file, including the ability to use an emulator.