Completed the authentication process for the application's user system.
This system allows users to create accounts, log out and in again, and delete the account. Deleting the account removes it from the Firebase collection as well in an irreversible(!) process.
Some style choices are still a bit unsatisfactory, and the code is a little too identical to the tutorial at the moment. I marked several TODOs for topics that I will pursue to change up my code and make sure it works well.
Styles on the three pages used for this practice (SignIn, Registration, Settings) are inconsistent with the test of the app.
Will need to further improve the account validation requirements once work on the UI of those features is implemented.
Removed the mock account fro Account.swift.