Added functionality to delete tasks from the user's list.
For an unknown reason, the only way I could get my SwiftUI list to update is by forcing a UI update, which removes the animation that came with removing items fom the list. I'm going to keep looking into how to get the update to occur without forcing it, but I wanted to get this committed before I start breaking things again.