Updated the navigation bar tabs to synchronize animations.
Implemented a testing zone called MinimalView. This will be used for when I need to break out small parts of my app t rest in an isolated environment without changing over to a separate project.
Removed the visibility boolean from the nav bar buttons, there was no good reason for it to be included.