Implemented the points mechanic into the Task system.
Whether points are added or not is dependent on whether the task was completed before it expired.
Added a maximum IP storage to accounts.
Added static function to Task to obfuscate the authModel from the views.
Added an 'X' that appears on a Task once it has expired. These tasks give no rewards when removed.