Began designing a framework of structures and views around the dungeon system.
Some content has been implemented, but I wouldn't consider this a finished system yet by any means.
Added a dungeonEndTime tracker to the user's Account.
Made the various CodingKeys used for en/decoding private.
Made a global variable for the main font of the app. Still need to choose a font and remember to use this.
Created a few files related to the incoming Dungeon system. Some of them have their decoding process started, but its mostly just inits and empty files at this point.
Changed the format of the rewards within the JSON file to better match JSON's expected format and return scheme.
Added the state tracker of the current dungeon to the DungeonState environment variable.