Implemented the visuals of the dungeon selection page.
The actual functionality isn't in yet, but that's next on the docket.
Set a bunch of environmental objects and other variables to private/private(set). It shouldn't have changed any of the existing functionality, but it seems to be a good style choice.
Hid the navigatoin logic into functions. Now, the developer needs to call a ChangeView function from the state managers rather than directly modifying the view's enum. I also moved around the default values into constants outside of the structs.
Set the app background to white until I figure out why some pages are forcing it to white. We'll leave this here until we make a dedicated session to figuring this out, which might not be before IDEAFest.