Some small cosmetic changed to the app's buttons
Modified app navigation code to only run if the destination is different from the current page.
Moved the origin of dungeonState to allow the active dungeon to persist between navigations.
Moved the dungeon's image into its own View, which will be reused in both the landing and adventure pages.
Narrowed the hitbox of a lot of buttons where I was using padding poorly.
Added directional arrows to the navigation bar tabs to indicate their purpose. I left in some code that was meant to get the arrows to change dynamically so that the overlapping visual clash went away, but I couldn't get it working. Maybe another day.