Created a file for simple data structures that are not supported by Swift.
The code within the file was AI generated and has not bug tested/finished yet, nor has it been implemented into the project yet.
Think of this commit as more of a sticky note reminder than a progress update. More will be coming in the next update, I assure you.