The user can now spend inspiration points and begin a dungeon.
The timer display on the remaining dungeon time is working, although I feel like I should add more words to it than just the time left.
Broke out most of the timer functions into their own struct DHMSTimer, which is being used by both the Task and Dungeon displays now.
Added the active dungeon Id to Firestore. This was a mistake to not include earlier.
Changed the dungeon time tracker in Account to a required variable instead of an optional. It should be invalid unless the dungeon Id is != 0 anyways, so this just makes Firestore easier to deal with.
The user is still able to change dungeons when a dungeon is active. This will need to change.
The user isn't able to mark a dungeon as complete yet, that will be coming next.
There is a framework for later dungeon features such as progression tracking and rewarding the user. This probably won't be worked on until after IdeaFest, though.
Speaking of which, a few IDEAFEST tags have been placed around to show changes that existing only for IdeaFest. These should be changed after the event.
Capitalized the names of a few functions.