Added image displays to the application.
Added 13 images to the Assets folder. 1 is for a profile image, while 12 are for the dungeons.
Dungeons now display an image during interactions with the system, such as selecting a dungeon to beginning an adventure.
The profile image shows a template empty-profile image. System to actually add/choose a profile image has not been implemented.
Stage progression in dungeons is not yet implemented, so only the first two images of each dungeon actually get used.
Moved the dungeon's name to the top of the image zone instead of the bottom so that the bench isn't being covered up.
Added 4 images to each of the 3 dungeons inside the JSON file.