Dungeon now returns the user to the proper dungeon if one is in-progress
Added some key accesses to allow testing on my iPhone. None of it looks sensitive, which is why it's being sent up.
Changed dungeonActiveId (Int) into activeDungeonName (string), which improves the ability to use it with dungeons. This also involved changing up a bunch of conditions to use strings instead.
Reduced the font size on the primary dungeon buttons.
Fixed the stutter issue on timers where they would display nothing on 0 seconds. Now, the timer skips over 0 seconds and shows the completion status instead.
Removed the non-login fallback code on the Task list.