Added custom authentication to the account creation process.
Created a new file to store the enums and functions used for authentication.
Added an error display to the Registration Page to inform the user when their attempt fails authentication.
Spread apart the text fields and buttons in the Sign In and Registration Pages. A following commit will provide more changes to their UI, but these are being added now.
The functions to authenticate are within the RegistrationView file for now, which is bad practice. They will be moved when I set up Password Changing in the Settings interface.