Added default behavior when the Display Name is left empty when creating an account.
Added lowercase forcing with emails when being passed in.
This does not change what the user is allowed to do, but makes sure that they can type it in whatever case they'd like without changing the outcome.
This was most likely already a feature with Firebase's email authentication, but it never hurts to make your own system on top.