Created a popout tab in the Profile view that will display the user's historical information.
The current plan is to display three pieces of data that have been collected over the account's lifetime: The IP collected, the number of tasks completed, and the number of dungeons completed.
The animation on popout and popin still need to be implemented.

Added a counter for lifetime IP in Account.
Refactored the NavBarArrows into a generic TabArrows subview that can support both vertical and horizontal stacking. This was then used for my new ProgressDetails tab as well.
Attempted to normalize the spacing on top of the buttons in the two task creation pages. The fix was hardcoded, and may cause an issue on different screen sizes. Although let's be honest, what won't, right? This app has only been tested on the 16 Pro for a reason.