Refined the SettingsView page to serve as a better host for account options.
Buttons have been temporarily added. The end goal is to create a list with navigation options.
Replaced the progress details tab opening button with a Stack onTab.
Added a 24 character limit to the Display Name. This will need to be enforced on the name changing feature as well.
Changed a button name from "Cancel" to "Return".